fix: decimal2Binary -> long int

+++ assembler.h
@@ -22,6 +22,6 @@ void firstPass(char**);
 void secondPass(char**);
 char* translateAInstruction(char*, int);
 char* translateCInstruction(char*);
-int decimal2Binary(int);
+long int decimal2Binary(int);
 bool isNumeric(const char*);
 char* leftPad(char*, char, int);
